break up project into multiple named files
