+++ lib/src/MeatPage.dart
@@ -1,0 +1,18 @@
+part of my_app;
+
+
+class PageTwo extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+        appBar: AppBar(
+          title: Text('Amys App'),
+        ),
+        body: Container(
+          padding: EdgeInsets.all(20.0),
+          child: Column(
+              crossAxisAlignment: CrossAxisAlignment.stretch,
+              children: [Text("page2")]),
+        ));
+  }
+}
