+++ pubspec.yaml
@@ -29,6 +29,7 @@ dependencies:
   firebase_core: ^0.4.0+9
   firebase_auth: ^0.14.0+5
   cloud_firestore: ^0.12.9+5
+  google_sign_in:
 
 
 dev_dependencies:
